rootfs/pwrstat/requirements.txt
flask
paho-mqtt<2.0.0
prometheus_client
ruamel.yaml
voluptuous
